🚚 Rename: 폴더명 수정(post > posting)
